% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agecomp.R
\name{agecomp}
\alias{agecomp}
\title{Generate age composition of brood year recruits}
\usage{
agecomp(ages, recruits, meanage, sdage)
}
\arguments{
\item{ages}{Numeric vector. All simulated age classes.}

\item{recruits}{Numeric. Number of recruits from that brood year.}

\item{meanage}{Numeric. Mean age of recruits.}

\item{sdage}{Numeric. Standard deviation of the age distribution.}
}
\value{
Numeric vector. Number of recruits in each age class.
}
\description{
Internal function to simulate the age distribution of recruits from a given brood year.
}
\examples{
\dontrun{
ages <- 2:6
agecomp(ages, recruits = 100, meanage = 4, sdage = 0.8)
}
}
\keyword{internal}
