% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{.get_review_years}
\alias{.get_review_years}
\title{Access review years from an ssi_run}
\usage{
.get_review_years(data)
}
\arguments{
\item{data}{An \code{ssi_run} object.}
}
\value{
Integer vector of review years.
}
\description{
Returns the sequence of review years used during the simulation.
This is saved into \code{ssi_run$parameters$review_years} by \code{\link[=run_scenarios]{run_scenarios()}}.
}
\keyword{internal}
