% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{.scenarios_from_ssirun}
\alias{.scenarios_from_ssirun}
\title{Standardize scenario labels and add IDs that match results list names}
\usage{
.scenarios_from_ssirun(data)
}
\arguments{
\item{data}{An \code{ssi_run} object}
}
\value{
Tibble with standardized scenario labels and keys
}
\description{
Used by Ohlberger figures. Adds scen_key matching the names in results lists.
}
\keyword{internal}
