% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-scenarios.R
\docType{data}
\name{scenarios}
\alias{scenarios}
\title{Scenarios list}
\format{
A tibble with rows × 11 columns:
\describe{
\item{ageT}{integer}
\item{sizeT}{integer}
\item{sexT}{integer}
\item{futureT}{character}
\item{trends}{character}
\item{sdsel}{double}
\item{maxsel}{double}
\item{selectivity}{factor}
\item{mgmt}{character}
\item{factorMSY}{factor}
\item{scenario_name}{character}
}
}
\source{
Generated from \code{data-raw/scenarios.R}.
}
\usage{
data(scenarios)
}
\description{
A tibble of scenarios from Ohlberger et al. 2025.
}
\references{
Ohlberger, J., Schindler, D. E., & Staton, B. A. (2025).
\emph{Accounting for salmon body size declines in fishery management can reduce conservation risks}.
Fish and Fisheries, 26, 113–130. \doi{10.1111/faf.12869}
}
\keyword{datasets}
