% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivity.R
\name{selectivity}
\alias{selectivity}
\title{Harvest selectivity based on fish size and mesh size}
\usage{
selectivity(size, meshsize, s)
}
\arguments{
\item{size}{Numeric. Fish size in millimeters.}

\item{meshsize}{Numeric. Stretch mesh size in inches.}

\item{s}{Numeric. Standard deviation of selectivity, typically \code{0.204}.}
}
\value{
Numeric. Proportion of fish of the given size retained by the net.
}
\description{
Internal function to calculate the proportion of fish retained by a gillnet
of a given mesh size, using the selectivity model described by Bromaghin (2005).
}
\examples{
selectivity(size = 550, meshsize = 5.0, s = 0.204)

}
\references{
Bromaghin, J. F. (2005). The statistical evaluation of relative gillnet selectivity
estimates and their application to the management of Alaskan sockeye salmon.
\emph{Alaska Fishery Research Bulletin}, 11(2), 258–266.
}
\keyword{internal}
