% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-imports.R
\docType{package}
\name{SizeShiftsImplicationsV2-package}
\alias{SizeShiftsImplicationsV2}
\alias{SizeShiftsImplicationsV2-package}
\title{Internal imports}
\description{
These imports are used across the package. They generate NAMESPACE entries
via roxygen and prevent R CMD check NOTES about “no visible global function
definition for ...”.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gholtgrieve.github/SizeShiftsImplicationsV2}
  \item Report bugs at \url{https://github.com/gholtgrieve/SizeShiftsImplicationsV2/issues}
}

}
\author{
\strong{Maintainer}: Gordon Holtgrieve \email{gholtgrieve@gmail.com}

}
\keyword{internal}
